finish api designing; start to work on different nodes
